requirements.txt
GitPython==3.1.40
python-dotenv==1.0.0
schedule==1.2.0

